Bug fixed when using multi channels devices and another minor issue (#86)
* Update index.js

Resolved a problem that occurred when using devices with multiple channels, like the Sonoff 4CH or the 2CH. There was an error saving the correct device id as key for the local cache. The result was that plugin try to create a new accessories with the classic error "Cannot add a bridged Accessory with the same UUID as another bridged Accessory".

* Update index.js

Resolved and error that may occur while adding a new Homebridge (so multiple accessories at the same time) in the iPhone Home app.